Tinkering with Priority Sorting
So I was digging around in the code and I figured I'd do my best to try to setup a way to sort tasks based upon their priority in Todoist.  It seems there are 4 levels of priority (colored flags) that you can set on a task (p1 - p4).  The values for each of these is reversed.  If you select P1 or "Priority 1" for your task (the red flag) then the value assigned in the data is "4" which means all of the ascending/descending needs to be reversed to sort things correctly.  

After I managed to get it to work on the first try, I realized that I need to figure out how to nest the priority sorting under the due date sorting so I can see what's most important for "today" by order of priority.  Otherwise, I end up with conflicting sorting.

Another possible solution could be to be able to choose how to sort.  "Sort first by date then priority, or sort first by priority then date."  Not sure how to nest those sorts into each other.